+++ frontend/src/components/@shared/UserProfileButton.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 import defaultUserImage from '../../assets/images/default_user.jpeg';
 import { ROUTE_PATH } from '../../constants/routes';
-import { useGetProfile } from '../../hooks/Profile/queries/profile';
+import { useGetProfile } from '../../hooks/@queries/profile';
 
 const UserProfileButton = () => {
   const { data: profile } = useGetProfile();
